requirements.txt
Flask>=2.2.0
pymongo>=4.0.0
flask-jwt-extended>=4.0.0
python-dotenv>=1.0.0